test: refactor integration suite
