+++ internal/testing/client.go
@@ -1,4 +1,4 @@
-package controllers
+package testing
 
 import (
 	"context"
